pyproject.toml
[tool.poetry]
name = "django-assessment"
version = "0.1.0"
description = "Senior Software Engineer Technical Assessment"
authors = ["Kiran Kumbhar <kiranpk189@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.15"
django-environ = "^0.11"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.8"
pytest-factoryboy = "^2.7"
factory-boy = "^3.3"
faker = "^24.0"
flake8 = "^7.0"
flake8-isort = "^6.1"
flake8-bugbear = "^24.0"
coverage = "^7.4"
pytest-cov = "^5.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=apps --cov-report=term-missing --cov-report=html -v"

[tool.coverage.run]
source = ["apps"]
omit = ["*/migrations/*", "*/tests/*"]

[tool.isort]
profile = "black"
known_django = "django"
known_rest_framework = "rest_framework"
sections = ["FUTURE","STDLIB","DJANGO","REST_FRAMEWORK","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
